students/k3341/laboratory_works/Mihaylov_Yriy/laboratiry_work_3/requirements.txt
asgiref==3.11.0
babel==2.17.0
backrefs==5.9
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
Django==6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-spectacular
drf-spectacular[sidecar]
ghp-import==2.1.0
idna==3.10
Jinja2==3.1.6
Markdown==3.9
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.20
mkdocs-material-extensions==1.3.1
oauthlib==3.3.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.4.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.3
pyyaml_env_tag==1.1
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0